# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MUFAZ Kitchen POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scheduled-orders:
    get:
      operationId: listScheduledOrders
      tags: [scheduled-orders]
      summary: List scheduled orders
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Scheduled orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduledOrder"
    post:
      operationId: createScheduledOrder
      tags: [scheduled-orders]
      summary: Create a scheduled order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduledOrderBody"
      responses:
        "201":
          description: Created scheduled order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledOrder"
  /scheduled-orders/{id}:
    get:
      operationId: getScheduledOrder
      tags: [scheduled-orders]
      summary: Get scheduled order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scheduled order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateScheduledOrder
      tags: [scheduled-orders]
      summary: Update a scheduled order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduledOrderBody"
      responses:
        "200":
          description: Updated scheduled order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteScheduledOrder
      tags: [scheduled-orders]
      summary: Cancel a scheduled order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scheduled-orders/{id}/convert:
    post:
      operationId: convertScheduledOrder
      tags: [scheduled-orders]
      summary: Convert scheduled order to a real order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Created order from scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "400":
          description: Already converted or cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scheduled-orders/daily-capacity:
    get:
      operationId: getDailyCapacity
      tags: [scheduled-orders]
      summary: Get daily order count and capacity for a date
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Capacity info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyCapacity"

  # ─── REPORTS ────────────────────────────────────────────────────────────────
  /reports/daily-sales:
    get:
      operationId: getDailySalesReport
      tags: [reports]
      summary: Daily sales summary with date range
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          required: true
          schema:
            type: string
        - name: dateTo
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily sales data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySalesReport"
  /reports/top-items:
    get:
      operationId: getTopItemsReport
      tags: [reports]
      summary: Top menu items by quantity sold
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          required: true
          schema:
            type: string
        - name: dateTo
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopItem"
  /reports/revenue-by-payment:
    get:
      operationId: getRevenueByPaymentReport
      tags: [reports]
      summary: Revenue breakdown by payment method
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          required: true
          schema:
            type: string
        - name: dateTo
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Revenue by payment method
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueByPayment"

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        itemPrice:
          type: string
        quantity:
          type: integer
        subtotal:
          type: string
      required:
        - id
        - orderId
        - itemName
        - itemPrice
        - quantity
        - subtotal
...
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        totalAmount:
          type: string
        paymentMethod:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - status
        - totalAmount
        - createdAt
        - updatedAt
        - items
...
    ScheduledOrderItem:
      type: object
      properties:
        menuItemId:
          type: integer
        itemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: string
      required:
        - menuItemId
        - itemName
        - quantity
        - unitPrice
    ScheduledOrder:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
        convertedOrderId:
          type: ["integer", "null"]
        items:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/ScheduledOrderItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - scheduledDate
        - scheduledTime
        - status
        - createdAt
        - updatedAt
    CreateScheduledOrderBody:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScheduledOrderItem"
      required:
        - scheduledDate
        - scheduledTime
        - items
    UpdateScheduledOrderBody:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        notes:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScheduledOrderItem"
    DailyCapacity:
      type: object
      properties:
        date:
          type: string
        currentCount:
          type: integer
        capacity:
          type: integer
        available:
          type: integer
      required:
        - date
        - currentCount
        - capacity
        - available

    # REPORTS
    DailySalesSummaryItem:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: string
        orderCount:
          type: integer
        avgOrderValue:
          type: string
      required:
        - date
        - revenue
        - orderCount
        - avgOrderValue
    DailySalesReport:
      type: object
      properties:
        summary:
          type: array
          items:
            $ref: "#/components/schemas/DailySalesSummaryItem"
        totalRevenue:
          type: string
        totalOrders:
          type: integer
      required:
        - summary
        - totalRevenue
        - totalOrders
    TopItem:
      type: object
      properties:
        menuItemId:
          type: integer
        itemName:
          type: string
        totalQuantity:
          type: integer
        totalRevenue:
          type: string
      required:
        - menuItemId
        - itemName
        - totalQuantity
        - totalRevenue
    RevenueByPayment:
      type: object
      properties:
        paymentMethod:
          type: string
        revenue:
          type: string
        orderCount:
          type: integer
      required:
        - paymentMethod
        - revenue
        - orderCount

    # SETTINGS
...