# lib/api-zod/src/generated/types/scheduledOrderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MUFAZ Kitchen POS API
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduledOrderItem {
  menuItemId: number;
  itemName: string;
  quantity: number;
  unitPrice: string;
}